# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Endurance API",
    "description": "Calculations for running, cycling, swimming, and triathlon.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cat-ranking/calculate": {
      "post": {
        "tags": [
          "Cycling Category"
        ],
        "summary": "Calculate cycling category from FTP",
        "description": "Determine cycling category from FTP w/kg.\n\nSystems:\n- usac: USA Cycling categories (sex-specific)\n- zwift: Zwift racing categories (A-D)\n- zwift-ce: Zwift Category Enforcement (A+ through E)",
        "operationId": "calculate_category_cat_ranking_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatRankingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatRankingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cat-ranking/thresholds": {
      "get": {
        "tags": [
          "Cycling Category"
        ],
        "summary": "Get category thresholds for a system",
        "description": "Return all category thresholds for a system.",
        "operationId": "get_thresholds_cat_ranking_thresholds_get",
        "parameters": [
          {
            "name": "system",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(usac|zwift|zwift-ce)$",
              "default": "usac",
              "title": "System"
            }
          },
          {
            "name": "sex",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[MF]$",
              "default": "M",
              "title": "Sex"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/critical-power/calculate": {
      "post": {
        "tags": [
          "Critical Power"
        ],
        "summary": "Calculate CP and W' from max efforts",
        "description": "Calculate Critical Power (CP) and W' using linear regression:\n\nWork = CP * time + W'\npower * time = CP * time + W'\npower = CP + W' / time\n\nUses 2+ maximal efforts of different durations (typically 2-20 min).",
        "operationId": "calculate_cp_critical_power_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriticalPowerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriticalPowerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/critical-power/predict": {
      "get": {
        "tags": [
          "Critical Power"
        ],
        "summary": "Predict power for a duration given CP/W'",
        "description": "Predict sustainable power for a given duration.",
        "operationId": "predict_power_critical_power_predict_get",
        "parameters": [
          {
            "name": "cp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Cp"
            }
          },
          {
            "name": "w_prime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "W Prime"
            }
          },
          {
            "name": "duration_sec",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Duration Sec"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/riegel/predict": {
      "post": {
        "tags": [
          "Race Prediction"
        ],
        "summary": "Predict race time using Riegel formula",
        "description": "Riegel formula: T2 = T1 * (D2/D1)^exponent\n\nDefault exponent 1.06 works well for running. \nCycling ~1.05, Swimming ~1.10.",
        "operationId": "predict_riegel_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiegelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiegelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/riegel/common": {
      "get": {
        "tags": [
          "Race Prediction"
        ],
        "summary": "Predict common race distances from one time",
        "description": "Predict 5k, 10k, half, marathon from any known distance.",
        "operationId": "common_distances_riegel_common_get",
        "parameters": [
          {
            "name": "distance_m",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Distance M"
            }
          },
          {
            "name": "time_sec",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Time Sec"
            }
          },
          {
            "name": "exponent",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.06,
              "title": "Exponent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/pace/convert": {
      "post": {
        "tags": [
          "Pace Conversion"
        ],
        "summary": "Convert between pace/speed units",
        "description": "Convert between pace and speed units.\n\nSupported units:\n- min_per_km: minutes per kilometer (running)\n- min_per_mi: minutes per mile (running)\n- kmh: kilometers per hour (cycling)\n- mph: miles per hour (cycling)\n- sec_per_100m: seconds per 100m (swimming)\n- sec_per_100y: seconds per 100 yards (swimming SCY)",
        "operationId": "convert_pace_convert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaceConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaceConvertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/pace/swim-convert": {
      "get": {
        "tags": [
          "Pace Conversion"
        ],
        "summary": "Convert swim paces between pool types",
        "description": "Convert swim pace between pool types.\n\nSCM = short course meters (25m)\nLCM = long course meters (50m)\nSCY = short course yards (25y)\n\nUses standard conversion factors.",
        "operationId": "swim_convert_pace_swim_convert_get",
        "parameters": [
          {
            "name": "pace_sec_per_100",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "title": "Pace Sec Per 100"
            }
          },
          {
            "name": "from_pool",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(scm|lcm|scy)$",
              "title": "From Pool"
            }
          },
          {
            "name": "to_pool",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(scm|lcm|scy)$",
              "title": "To Pool"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatRankingRequest": {
        "properties": {
          "ftp_watts": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Ftp Watts",
            "description": "FTP in watts"
          },
          "weight_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Weight Kg",
            "description": "Body weight in kg"
          },
          "sex": {
            "type": "string",
            "pattern": "^[MF]$",
            "title": "Sex",
            "description": "Sex: M or F",
            "default": "M"
          },
          "system": {
            "type": "string",
            "title": "System",
            "description": "Ranking system: usac, zwift, zwift-ce",
            "default": "usac"
          }
        },
        "type": "object",
        "required": [
          "ftp_watts",
          "weight_kg"
        ],
        "title": "CatRankingRequest"
      },
      "CatRankingResponse": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Assigned category"
          },
          "ftp_wkg": {
            "type": "number",
            "title": "Ftp Wkg",
            "description": "FTP in watts/kg"
          },
          "thresholds": {
            "additionalProperties": true,
            "type": "object",
            "title": "Thresholds",
            "description": "All thresholds for reference"
          },
          "next_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Category",
            "description": "Next category up"
          },
          "watts_to_next": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Watts To Next",
            "description": "Watts needed for next category"
          }
        },
        "type": "object",
        "required": [
          "category",
          "ftp_wkg",
          "thresholds"
        ],
        "title": "CatRankingResponse"
      },
      "CriticalPowerRequest": {
        "properties": {
          "efforts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Efforts",
            "description": "List of {duration_sec, power_watts} for 2+ max efforts"
          }
        },
        "type": "object",
        "required": [
          "efforts"
        ],
        "title": "CriticalPowerRequest"
      },
      "CriticalPowerResponse": {
        "properties": {
          "cp": {
            "type": "number",
            "title": "Cp",
            "description": "Critical Power (watts)"
          },
          "w_prime": {
            "type": "number",
            "title": "W Prime",
            "description": "W' (joules)"
          },
          "r_squared": {
            "type": "number",
            "title": "R Squared",
            "description": "Fit quality"
          }
        },
        "type": "object",
        "required": [
          "cp",
          "w_prime",
          "r_squared"
        ],
        "title": "CriticalPowerResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaceConvertRequest": {
        "properties": {
          "value": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Value",
            "description": "Input value"
          },
          "from_unit": {
            "type": "string",
            "title": "From Unit",
            "description": "Source unit: min_per_km, min_per_mi, kmh, mph, sec_per_100m, sec_per_100y"
          },
          "to_unit": {
            "type": "string",
            "title": "To Unit",
            "description": "Target unit"
          }
        },
        "type": "object",
        "required": [
          "value",
          "from_unit",
          "to_unit"
        ],
        "title": "PaceConvertRequest"
      },
      "PaceConvertResponse": {
        "properties": {
          "result": {
            "type": "number",
            "title": "Result"
          },
          "from_unit": {
            "type": "string",
            "title": "From Unit"
          },
          "to_unit": {
            "type": "string",
            "title": "To Unit"
          },
          "input_value": {
            "type": "number",
            "title": "Input Value"
          }
        },
        "type": "object",
        "required": [
          "result",
          "from_unit",
          "to_unit",
          "input_value"
        ],
        "title": "PaceConvertResponse"
      },
...
      "RiegelRequest": {
        "properties": {
          "known_distance_m": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Known Distance M",
            "description": "Known race distance in meters"
          },
          "known_time_sec": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Known Time Sec",
            "description": "Known race time in seconds"
          },
          "target_distance_m": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Target Distance M",
            "description": "Target distance in meters"
          },
          "exponent": {
            "type": "number",
            "title": "Exponent",
            "description": "Riegel exponent (default 1.06)",
            "default": 1.06
          }
        },
        "type": "object",
        "required": [
          "known_distance_m",
          "known_time_sec",
          "target_distance_m"
        ],
        "title": "RiegelRequest"
      },
      "RiegelResponse": {
        "properties": {
          "predicted_time_sec": {
            "type": "number",
            "title": "Predicted Time Sec"
          },
          "known_distance_m": {
            "type": "integer",
            "title": "Known Distance M"
          },
          "known_time_sec": {
            "type": "number",
            "title": "Known Time Sec"
          },
          "target_distance_m": {
            "type": "integer",
            "title": "Target Distance M"
          },
          "exponent": {
            "type": "number",
            "title": "Exponent"
          }
        },
        "type": "object",
        "required": [
          "predicted_time_sec",
          "known_distance_m",
          "known_time_sec",
          "target_distance_m",
          "exponent"
        ],
        "title": "RiegelResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}